Add support for P2P Mode on find images.
With the P2P Mode is possible to retrieve
the container's images on all remotes
configured and to use the image with the
more updated creation date.

The default mode is the normal mode, in
this case, a remote name string is
required.

The LXCExecutor object has now the methods SetP2PMode()
and GetP2PMode().

The mode could be set from cli with --p2p-mode option
or through the .lxd-compose.yaml under the genaral
options.