executor: Add LocalDisable option
Now it's possible to disable the use of local remote that is added
automatically from LXD API.
This could be help Mottainai Agent on control remote LXD instance from a
container where LXD is not running.

This option could be set from config with the new general option
`lxd_local_disable`.