Add `wait_ip` attribute on node
Using LXD on node where the hardisk is slow or under heavy load
could generates delay on ip assign and/or iface setup.
This is critical for CD/CI tasks. We add an `wait_ip` attribute
where the value is the timeout in seconds that is used by
lxd-compose to waiting for ip address.

So, when `wait_ip` is greather then 0 and the ip address is
been assigned a new emit event "container_ip" is generated.