render: Using Metadata name tpl for templates
In order to cleanup the use of the templates
we use `tpl` string as metadata name. This means
that `include` of templates must be with prefix `tpl`.

In addition, to help users with troubleshooting of
the render engine is added the env variable
LXD_COMPOSE_HELM_DEBUG that permit to see on stdout
the rendered YAML before the processing and catch
easily issues with indentations or similar.